feat: update popover direction
