+++ apps/website/docs/components/Actions/popover/Right.tsx
@@ -1,0 +1,37 @@
+import {
+  button,
+  buttonGhost,
+  clickReaction,
+  clickReactionInput,
+  popover,
+  popoverContent,
+  popoverRight,
+} from "@idealjs/camphora-styled";
+import clsx from "clsx";
+import React from "react";
+
+const id = "popoverRight";
+
+const Popover = () => {
+  return (
+    <div
+      style={{
+        height: "100%",
+        display: "flex",
+        alignItems: "center",
+        justifyContent: "center",
+      }}
+    >
+      <label
+        htmlFor={id}
+        className={clsx(clickReaction, popover, button, buttonGhost)}
+      >
+        <input id={id} type="checkbox" className={clickReactionInput} />
+        <div className={clsx(popoverContent, popoverRight)}>hello popover</div>
+        open
+      </label>
+    </div>
+  );
+};
+
+export default Popover;
