+++ packages/camphora-styled/src/previewBlock.css.ts
@@ -6,5 +6,5 @@ export const previewBlock = style({
   padding: "16px",
   backgroundSize: "5px 5px",
   backgroundImage: `radial-gradient(${vars.colors.base.content} 0.5px ,${vars.colors.base[200]} 0.5px)`,
-  minHeight: "300px",
+  height: "300px",
 });
